Redesign the LoggerSettings class. Remove the LoggerSettings and put the PII and logcat properties into the Logger class. Add the static public API to get the setting variables.
