+++ common/src/main/java/com/microsoft/identity/common/internal/logging/Logger.java
@@ -18,6 +18,11 @@ public final class Logger {
     private LogLevel mLogLevel  = LogLevel.VERBOSE;
     private AtomicReference<ILoggerCallback> mExternalLogger = new AtomicReference<>(null);
 
+    // Disable to log PII by default.
+    private static boolean mAllowPii = false;
+    // Disable to Logcat logging by default.
+    private static boolean mAllowLogcat = false;
+
     /**
      * Enum class for LogLevel that the sdk recognizes.
      */
@@ -45,6 +50,39 @@ public final class Logger {
      */
     public static Logger getInstance() {
         return INSTANCE;
+    }
+
+    /**
+     *  Enable/Disable log message with PII (personal identifiable information) info.
+     *  By default, the SDK doesn't log any PII.
+     *
+     * @param allowPii True if enabling PII info to be logged, false otherwise.
+     */
+    public static void setAllowPii(final boolean allowPii) {
+        mAllowPii = allowPii;
+    }
+
+    /**
+     * Enable/Disable the Android logcat logging. By default, the sdk disables it.
+     *
+     * @param allowLogcat True if enabling the logcat logging, false otherwise.
+     */
+    public static void setAllowLogcat(final boolean allowLogcat) {
+        mAllowLogcat = allowLogcat;
+    }
+
+    /**
+     * Get if log PII is enabled.
+     */
+    public static boolean getAllowPii() {
+        return mAllowPii;
+    }
+
+    /**
+     * Get if logcat is enabled.
+     */
+    public static boolean getAllowLogcat() {
+        return mAllowLogcat;
     }
 
     /**
@@ -142,8 +180,8 @@ public final class Logger {
             return;
         }
 
-        // Developer turns off PII logging, if the log message contains any PII, we should not send it.
-        if (!LoggerSettings.getAllowPii() && containsPII) {
+        // Developer turns off PII logging, if the log meLoggerSettingssage contains any PII, we should not send it.
+        if (!mAllowPii && containsPII) {
             return;
         }
 
@@ -151,7 +189,7 @@ public final class Logger {
         final String logMessage = formatMessage(correlationID, message, throwable);
 
         // Send logs into Logcat.
-        if (LoggerSettings.getAllowLogcat()) {
+        if (mAllowLogcat) {
             sendLogcatLogs(tag, logLevel, logMessage.toString());
         }
